mysite_project/requirements.txt
channels==2.1.6f
defusedxml==0.5.0
Django==2.1.5
django-debug-toolbar==1.10.1
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
djangorestframework-xml==1.3.0
gunicorn==19.9.0
Pillow==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.7
channels_redis==2.3.3

